Fix : double free detected bug

+++ Phase2/headers.h
@@ -25,6 +25,8 @@ typedef short bool;
 #define MSG_PROC_DOWN_KEY 'D'
 #define MSG_PROC_UP_KEY 'U'
 
+bool terminateSignalSent = 0;
+
 enum Algorithms
 {
     HPF,
@@ -39,8 +41,8 @@ struct processInfo
     int arrivalTime;
     int runTime;
     int priority;
-    bool isRunning;
     int pid;
+    bool isRunning;
     int remainingTime;
     int finishTime;
     int startTime;
@@ -133,8 +135,10 @@ void initClk()
 void destroyClk(bool terminateAll)
 {
     shmdt(shmaddr);
-    if (terminateAll)
+    if (terminateAll && !terminateSignalSent)
     {
+        printf("Call from destory clk \n");
+        terminateSignalSent = 1;
         killpg(getpgrp(), SIGINT);
     }
 }
